Convert chainable generators to ES Modules
