--- lib/chainables/chunk.js
@@ -1,25 +1,0 @@
-const { safeIterate, iterate } = require("../base-iterators");
-const take = require("./take");
-
-/**
- * Takes values from the supplied Iterable and yields Arrays of length n. If the Iterable object is finite,
- * the last chunk yielded will contain as many items as possible.
- *
- * @param {Iterable} it Any iterable object
- * @param {number} n An integer indicating the maximum size of the Arrays to be yielded.
- */
-function* chunk(it, n = 1) {
-	let source = iterate(it);
-
-	try {
-		while (true) {
-			let next = [...take(safeIterate(source), n)];
-			if (!next.length) return;
-			yield next;
-		}
-	} finally {
-		source.return();
-	}
-}
-
-module.exports = chunk;
