+++ lib/chainables/compact.mjs
@@ -1,5 +1,5 @@
-const { identity } = require("../utils");
-const filter = require("./filter");
+import { identity } from "../utils";
+import filter from "./filter";
 
 /**
  * Takes values from the supplied Iterable and yields only truthy values. The falsey values `false`, `null`,
@@ -7,8 +7,6 @@ const filter = require("./filter");
  *
  * @param {Iterable} it Any iterable object
  */
-function* compact(it) {
-	yield* filter(it, identity);
+export default function* compact(it) {
+    yield* filter(it, identity);
 }
-
-module.exports = compact;
