+++ lib/chainables/concat.mjs
@@ -1,4 +1,4 @@
-const { isIterable } = require("../base-iterators");
+import { isIterable } from "../base-iterators/index";
 
 /**
  * Takes any number of parameters and sequentially yields all values from them. If the given object is
@@ -7,14 +7,12 @@ const { isIterable } = require("../base-iterators");
  *
  * @param {...(Iterable|*)} them Any objects or values
  */
-function* concat(...them) {
-	for (let it of them) {
-		if (isIterable(it)) {
-			yield* it;
-		} else {
-			yield it;
-		}
-	}
+export default function* concat(...them) {
+    for (let it of them) {
+        if (isIterable(it)) {
+            yield* it;
+        } else {
+            yield it;
+        }
+    }
 }
-
-module.exports = concat;
