+++ lib/chainables/difference.mjs
@@ -1,5 +1,5 @@
-const { iterate } = require("../base-iterators");
-const flatten = require("./flatten");
+import { iterate } from "../base-iterators/index";
+import flatten from "./flatten";
 
 /**
  * Exclude any items from the given Iterable that are also present in the provided values. Any Iterable
@@ -9,15 +9,13 @@ const flatten = require("./flatten");
  * @param {Iterable} values One or more arrays or finitely iterable objects containing the set of values
  *         to be excluded by the generator.
  */
-function* difference(it, ...values) {
-	let source = iterate(it);
-	let exclusions = [...flatten(values, 1)];
+export default function* difference(it, ...values) {
+    let source = iterate(it);
+    let exclusions = [...flatten(values, 1)];
 
-	for (let value of source) {
-		if (!exclusions.includes(value)) {
-			yield value;
-		}
-	}
+    for (let value of source) {
+        if (!exclusions.includes(value)) {
+            yield value;
+        }
+    }
 }
-
-module.exports = difference;
