+++ lib/chainables/differenceBy.mjs
@@ -1,5 +1,5 @@
-const { iterate } = require("../base-iterators");
-const { identity } = require("../utils");
+import { iterate } from "../base-iterators/index";
+import { identity } from "../utils";
 
 /**
  * Exclude any items from the given Iterable that are also present in the provided values, where each value
@@ -11,15 +11,13 @@ const { identity } = require("../utils");
  *         to be excluded by the generator.
  * @param mapFn
  */
-function* differenceBy(it, values, mapFn = identity) {
-	let source = iterate(it);
-	let exclusions = Array.from(values, mapFn);
+export default function* differenceBy(it, values, mapFn = identity) {
+    let source = iterate(it);
+    let exclusions = Array.from(values, mapFn);
 
-	for (let value of source) {
-		if (!exclusions.includes(mapFn(value))) {
-			yield value;
-		}
-	}
+    for (let value of source) {
+        if (!exclusions.includes(mapFn(value))) {
+            yield value;
+        }
+    }
 }
-
-module.exports = differenceBy;
