--- lib/chainables/filter.js
@@ -1,18 +1,0 @@
-const { iterate } = require("../base-iterators");
-const { identity } = require("../utils");
-
-/**
- *
- * @param {Iterable} it Any iterable object
- * @param predicate
- */
-function* filter(it, predicate = identity) {
-	let i = 0;
-	for (let value of iterate(it)) {
-		if (predicate(value, i++)) {
-			yield value;
-		}
-	}
-}
-
-module.exports = filter;
