--- lib/chainables/forEach.js
@@ -1,17 +1,0 @@
-const { iterate } = require("../base-iterators");
-const { identity } = require("../utils");
-
-/**
- *
- * @param {Iterable} it Any iterable object
- * @param fn
- */
-function* forEach(it, fn = identity) {
-	let index = 0;
-	for (let value of iterate(it)) {
-		fn(value, index++);
-		yield value;
-	}
-}
-
-module.exports = forEach;
