+++ lib/chainables/index.mjs
@@ -3,28 +3,28 @@
  * @author Lachlan Hunt <lachlan.hunt@lachy.id.au>
  */
 
-export const chunk = require("./chunk");
-export const compact = require("./compact");
-export const concat = require("./concat");
-export const dedupe = require("./dedupe");
-export const difference = require("./difference");
-export const differenceBy = require("./differenceBy");
-export const differenceWith = require("./differenceWith");
-export const drop = require("./drop");
-export const dropUntil = require("./dropUntil");
-export const dropWhile = require("./dropWhile");
-export const everyNth = require("./everyNth");
-export const filter = require("./filter");
-export const flatten = require("./flatten");
-export const forEach = require("./forEach");
-export const fuzz = require("./fuzz");
-export const interlace = require("./interlace");
-export const link = require("./link");
-export const map = require("./map");
-export const modulo = require("./modulo");
-export const pluck = require("./pluck");
-export const take = require("./take");
-export const takeUntil = require("./takeUntil");
-export const takeWhile = require("./takeWhile");
-export const unzip = require("./unzip");
-export const zip = require("./zip");
+export { default as chunk } from "./chunk";
+export { default as compact } from "./compact";
+export { default as concat } from "./concat";
+export { default as dedupe } from "./dedupe";
+export { default as difference } from "./difference";
+export { default as differenceBy } from "./differenceBy";
+export { default as differenceWith } from "./differenceWith";
+export { default as drop } from "./drop";
+export { default as dropUntil } from "./dropUntil";
+export { default as dropWhile } from "./dropWhile";
+export { default as everyNth } from "./everyNth";
+export { default as filter } from "./filter";
+export { default as flatten } from "./flatten";
+export { default as forEach } from "./forEach";
+export { default as fuzz } from "./fuzz";
+export { default as interlace } from "./interlace";
+export { default as link } from "./link";
+export { default as map } from "./map";
+export { default as modulo } from "./modulo";
+export { default as pluck } from "./pluck";
+export { default as take } from "./take";
+export { default as takeUntil } from "./takeUntil";
+export { default as takeWhile } from "./takeWhile";
+export { default as unzip } from "./unzip";
+export { default as zip } from "./zip";
