+++ lib/chainables/modulo.mjs
@@ -1,4 +1,4 @@
-const { iterate } = require("../base-iterators");
+import { iterate } from "../base-iterators/index";
 
 /**
  * For dividends generted by `it`, yields the value mod `n` based on Euclidian division.
@@ -8,10 +8,8 @@ const { iterate } = require("../base-iterators");
  * @param  {Number}    n  The divisor
  * @return {Generator}
  */
-function* modulo(it, n) {
-	for (let value of iterate(it)) {
-		yield (value % n + Math.abs(n)) % n;
-	}
+export default function* modulo(it, n) {
+    for (let value of iterate(it)) {
+        yield (value % n + Math.abs(n)) % n;
+    }
 }
-
-module.exports = modulo;
