--- lib/chainables/take.js
@@ -1,18 +1,0 @@
-const { iterate } = require("../base-iterators");
-
-/**
- *
- * @param {Iterable} it Any iterable object
- * @param n
- */
-function* take(it, n = 1) {
-	let count = 0;
-	if (n <= 0) return;
-
-	for (let value of iterate(it)) {
-		yield value;
-		if (++count >= n) return;
-	}
-}
-
-module.exports = take;
