+++ lib/chainables/zip.mjs
@@ -1,0 +1,16 @@
+import { iterate } from "../base-iterators/index";
+
+/**
+ *
+ * @param them
+ */
+export default function* zip(...them) {
+    if (!!them.length) {
+        let iterators = them.map(iterate);
+        while (true) {
+            let values = iterators.map(iterator => iterator.next());
+            if (values.some(value => value.done)) return;
+            yield values.map(value => value.value);
+        }
+    }
+}
